requirements.txt
boto3==1.35.40
celery==5.4.0
django==5.1.3
gunicorn==23.0.0
pycurl==7.45.3
python-decouple==3.8
python-telegram-bot==21.7



